feat: streamline Docker registry setup in CI workflow (#1)
* feat: streamline Docker registry setup in CI workflow

Refactor the CI workflow to utilize an Ansible playbook for 
setting up a private Docker registry. 

* feat: add logging for with-registry-auth flag in deploy command

Adds a log message to indicate when the with-registry-auth flag is 
included in the Docker stack deploy command. This change improves 
visibility into the deployment process and helps with debugging 
by clearly showing the command's configuration.